requirements.txt
streamlit
google-cloud-firestore
watchdog
pandas
numpy
pandas>=2.2.2
matplotlib>=3.9.0
transformers>=4.40.0
torch>=2.3.0
wordcloud
scikit-learn
scipy
matplotlib
nltk
plotly
authlib
google-cloud-storage